// apps/terminal/ui/base/input/src/types.ts
type InputFieldOptionsBase = Readonly<{
  readonly fieldId: string;
  readonly testID: string;
  readonly accessibilityLabel?: string;
  readonly editable?: boolean;
  readonly initialValue?: string;
  readonly initialSelection?: PrimitiveInputSelection;
  readonly maxLength?: number;
  readonly secureTextEntry?: boolean;
}>;

export type InputFieldOptions =
  | (InputFieldOptionsBase &
      Readonly<{
        readonly keyboardKind: 'system';
        readonly layout?: never;
      }>)
  | (InputFieldOptionsBase &
      Readonly<{
        readonly keyboardKind: 'virtual';
        readonly layout: KeyboardLayout;
      }>);

export type InputFieldResult = Readonly<{
  readonly inputProps: PrimitiveInputProps;
  readonly captureInputSnapshot: () => InputSnapshot;
  readonly focus: () => void;
  readonly blur: () => void;
  readonly complete: () => void;
}>;